Wireup tracing for Express and HTTPS
